# next.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inbound API v2",
    "version": "2.0.0",
    "description": "The official API documentation for Inbound - Email API Service. Manage domains, send emails, handle webhooks, and more.",
    "contact": {
      "name": "Inbound Support",
      "url": "https://inbound.new/support",
      "email": "support@inbound.new"
    },
    "license": {
      "name": "Commercial",
      "url": "https://inbound.new/terms"
    }
  },
  "servers": [
    {
      "url": "https://inbound.new/api/v2",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api/v2",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "API Key authentication. Use 'Bearer YOUR_API_KEY'"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for session-based authentication"
      }
    },
...
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                },
                "details": {
                  "type": "string",
                  "example": "Additional error details"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred. Please try again later."
                },
                "code": {
                  "type": "string",
                  "example": "500"
                },
                "details": {
                  "type": "string",
                  "example": "Additional error details"
                }
              }
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/v2/onboarding/webhook": {
      "post": {
        "operationId": "post-v2-onboarding-webhook",
        "summary": "Handle onboarding webhook",
        "description": "Process webhook events for onboarding demo emails. Updates demo status when replies are received.",
        "tags": [
          "Onboarding"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  }
}